adding support for simple classes, added call method mangling, removed lists for the time being

+++ examples/dgemm.h
@@ -2,7 +2,7 @@ typedef struct p_class p_class;
 
 struct p_class{
 int attr_1;
-double attr_2;
+float attr_2;
 void (*_Z8method_2prmt_matprmt_matprmt_mat)(p_class *self, struct prmt_mat * A, struct prmt_mat * B, struct prmt_mat * C);
 };
